requirements.txt
anthropic>=0.39.0
flask>=3.0.0
google-genai>=1.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
pytest>=7.0.0
reportlab>=4.0.0
